Complete refactor: notes as follows
- The update endpoint no longer points to an actualy download link, just the clone hero main page
- Instead I've changed the functionality to crawl the download page and look for the appropriate download link & download it from there. Kind of brute forcey but it works
- Folder/file names are set once in the establishFileName function, easier for page crawling & file moving.
- Mega no longer used (though could be an option because it's still on the download page)
- Some error formatting (for debugging curiosity)

One issue persists: The unarr library returns "unarr: Error Read" for some reason or another, that I can't seem to figure out. Maybe the clonehero site is sending weirdly zipped files? I'm at a loss. Modifying the library to not panic seems to fix the issue and the game runs fine, but that's hard a solution lol.